Update MinIMU-9-test.py to run with VPython 7 and python3
Before this commit, the python script was only compatible with older versions of Vpython (5.74) and python 2.7.

This modified script can be run with latest pip version of Vpython and python 3.X. It was tested in a virtualenv on Linux:
Python: 3.9.4
vpython: 7.6.1

This new version does basically the same thing than the previous one but in the navigator.